move TransitMap and Polyline
